Finishing App

+++ src/main/java/com/rangotech/springsecurityapp/persistence/entity/Cart.java
@@ -16,7 +16,7 @@ public class Cart {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer cartId;
+    private Long cartId;
 
     @OneToOne()
     @JoinColumn(name = "user_id")
